Conflicts:
	addons/.cproject
	libs/openFrameworks/.cproject
	libs/openFrameworks/.project
	libs/openFrameworks/3d/ofVbo.cpp
	libs/openFrameworks/3d/ofVbo.h
	libs/openFrameworks/3d/ofVboMesh.cpp
	libs/openFrameworks/3d/ofVboMesh.h
	libs/openFrameworks/3d/ofVertexData.h
	libs/openFrameworks/graphics/ofGraphics.cpp
	libs/openFrameworks/graphics/ofShape.cpp
	libs/openFrameworks/graphics/ofTrueTypeFont.cpp
	libs/openFrameworks/graphics/ofTrueTypeFont.h

+++ libs/openFrameworks/3d/ofVertexData.cpp
@@ -181,13 +181,13 @@ void ofVertexData::addIndex(int i){
 }
 
 //--------------------------------------------------------------
-void ofVertexData::addIndices(const vector<GLuint>& inds){
+void ofVertexData::addIndices(const vector<GLubyte>& inds){
 	indices.insert(indices.end(),inds.begin(),inds.end());
 	bIndicesChanged = true;
 }
 
 //--------------------------------------------------------------
-void ofVertexData::addIndices(const GLuint* inds, int amt){
+void ofVertexData::addIndices(const GLubyte* inds, int amt){
 	for (int i = 0; i < amt;i++){
 		addIndex(inds[i]);
 	}
@@ -278,7 +278,7 @@ float* ofVertexData::getTexCoordsPointer(){
 }
 
 //--------------------------------------------------------------
-GLuint* ofVertexData::getIndexPointer(){
+GLubyte* ofVertexData::getIndexPointer(){
 	return &indices[0];
 }
 
@@ -399,6 +399,6 @@ void ofVertexData::setupIndicesAuto(){
 	bIndicesChanged = true;
 	indices.clear();
 	for(int i = 0; i < (int)vertices.size();i++){
-		indices.push_back((GLuint)i);
+		indices.push_back((GLubyte)i);
 	}
 }
